Phase 2 slice 5a: fractional zoom plumbing, analytical FitBounds
MapState.Zoom, Cfg.InitialZoom, OnZoomChange, Projector.Zoom,
panState.StartZoom flip uint32→float64. viewport gains TileZ (integer
floor) + TileScale (2^(Z-TileZ)); drawTiles fetches at TileZ, draws at
ceil(TileSize*TileScale) to suppress seams.

projection adds ProjectF / UnprojectF / WorldSizeF; integer variants
delegate. WorldSizeF / UnprojectF guard NaN/±Inf.

FitBounds rewritten analytical: min(log2(availW/dx0), log2(availH/dy0))
clamped [0, maxZoomF]. Padding-consumes-canvas → Zoom=0 via NaN
propagation through math.Min.

clampZoom helper consolidates NaN/±Inf/negative/ceiling guard across
every zoom write path. isFinite helper unifies finite checks; isFiniteF32
and finitePositive now delegate. zoomLabel formats integer zoom as
"12", fractional as "12.4"; shared by coord readout, zoom chip, and
A11YDescription.

Wheel + keyboard +/- keep integer deltas; fractional state reached
via SetZoom / SetView / FitBounds only.

Tests: float projection round-trip at fractional z, WorldSizeF parity
+ non-finite collapse, UnprojectF NaN-point collapse, viewport
TileZ/TileScale at fractional + max zoom + integer, drawTiles emits
ceil'd tile size, analytical FitBounds (fractional / integer-exact /
max-zoom clamp / degenerate axis / padding-consumed), clampZoom table,
SetZoom/SetView NaN-zoom collapse, SetZoom fractional round-trip,
zoomToward fractional-to-fractional, metersPerPixel continuity,
zoomLabel branches.

+++ examples/full-map/main.go
@@ -20,7 +20,7 @@ var src = tile.OSMWithUserAgent(
 type city struct {
 	Name   string
 	Center projection.LatLng
-	Zoom   uint32
+	Zoom   float64
 }
 
 var cities = []city{
